pyproject.toml
[tool.poetry]
name = "spotify-organizer"
version = "0.1.0"
description = "Organize songs in Spotifyy by tempo, time signature etc"
authors = ["Ipek Kantarci <ipke94@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
spotipy = "2.24.0"
python-dotenv = "^1.0.1"
tzdata = "^2024.1"
